text-extraction-func/requirements.txt
azure-functions
azure-ai-formrecognizer>=3.2.0
azure-core>=1.23.0
azure-identity>=1.12.0
PyMuPDF>=1.21.1
opencv-python>=4.5.0
opencv-python-headless==4.11.0.86
pillow>=9.0.0
numpy>=1.20.0
requests>=2.28.0
python-dotenv